YABI-333: STDERR being lost on epic jobs
Passing in stderr and stdout files to qsub directly.
This is the way the old BE was doing it and I think it simplifies everything.

STDOUT.txt and STDERR.txt will end up in the workingdir without us having to do anything special about it.
We also don't rely on having lcopy on any of the BEs.