Don't allow previous test run's stageout dir reuse.
If the same bamboo agent runs the tests the previous stageout dirs are
reused. Since we know allow symlinks, in stageout dirs we have some
broken symlinks when later tests with same workflow names are run.

Ideally, we would like to delete the dirs before we start running the
tests, but the bamboo user isn't allowed to do that.
So, as a workaround I changed the workflow name to have the current time
in the name as well.
Not ideal, but efficient...