Workflow reuse fails if a tool is renamed
When adding the tool to a workflow we're going to look it up just by
toolId.
Before the tool/tool description split it was looked up by name, at
the split it has been changed to being looked up by name and id.

I don't see the reason why we should not look it up just by tool id.
This will solve the problem of the possible tool name changes.

Fixes #3.